pyproject.toml
[project]
name = "jigsaw"
version = "0.1.5"
description = "This is my codebase for Jigsaw competition."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clean-text>=0.6.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "iterative-stratification>=0.1.9",
    "joblib>=1.5.2",
    "kaggle>=1.7.4.5",
    "kagglehub>=0.3.13",
    "matplotlib>=3.10.6",
    "mlflow>=3.4.0",
    "nlpaug",
    "numpy>1.26,<2.0",
    "pandas>=2.3.2",
    "pronouncing>=0.2.0",
    "python-box>=7.3.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.1",
    "tensorboard>=2.20.0",
    "tiktoken>=0.11.0",
    "torch>=2.7.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "transformers[torch]>=4.56.2",
    "typeguard>=4.4.4",
    "types-pyyaml>=6.0.12.20250915",
    "unidecode>=1.4.0",
    "wordcloud>=1.9.4",
]

[tool.uv.sources]
nlpaug = { git = "https://github.com/makcedward/nlpaug.git" }
